requirements/local.txt
-r base.txt

# Testing
pytest
pytest-django
pre-commit
coverage
